Implement course details page, move model interfaces to shared, #57
