+++ api/src/security/JwtUtils.ts
@@ -1,4 +1,4 @@
-import {IUser} from '../models/IUser';
+import {IUser} from '../../../shared/models/IUser';
 import {sign} from 'jsonwebtoken';
 import config from '../config/main';
