+++ app/webFrontend/src/app/course/course-detail/course-detail.component.html
@@ -1,21 +1,25 @@
-<md-card>
-
-    <p>
-    {{name}}
-</p>
-
-<p>
-  Your current level: Lecture 1
-</p>
-
-<ul>
-  <li>Lecture 1</li>
-  <li>Lecture 2</li>
-</ul>
-
-    <iframe src="http://node.t-sauer.de/" width="100%" height="500"></iframe>
-
-<button (click)="apply()">
-    apply
-</button>
+<md-card *ngIf="course">
+  <md-card-header>
+    <md-card-title>{{course.name}}</md-card-title>
+    <md-card-subtitle>{{course.description}}</md-card-subtitle>
+  </md-card-header>
+  <md-card-content>
+    <md-card *ngFor="let lecture of course.lectures;">
+      <md-card-header>
+        <md-card-title>{{lecture.name}}</md-card-title>
+        <md-card-subtitle>{{lecture.description}}</md-card-subtitle>
+      </md-card-header>
+      <md-card-content>
+        <md-card *ngFor="let unit of lecture.units;">
+          <md-card-header>
+            <md-card-title>{{unit._id}}</md-card-title>
+          </md-card-header>
+          <md-card-content>
+            Filename: {{unit.fileName}}
+            Filepath: {{unit.filePath}}
+          </md-card-content>
+        </md-card>
+      </md-card-content>
+    </md-card>
+  </md-card-content>
 </md-card>
