+++ app/webFrontend/src/app/dashboard/dashboard.component.html
@@ -21,7 +21,7 @@ Available Courses
             </md-card-content>
             <md-card-actions>
                 <button md-button (click)="apply(course._id)">Apply</button>
-                <button md-button (click)="goToInfo(course.name)">Details</button>
+                <button md-button (click)="goToInfo(course._id)">Details</button>
                 <button md-button (click)="editCourse(course._id)" *ngIf="this.userService.isTeacher()">Edit</button>
 
             </md-card-actions>
