+++ app/webFrontend/src/app/models/course.ts
@@ -1,4 +1,13 @@
-export class Course {
+import {ICourse} from "../../../../../shared/models/ICourse";
+import {IUser} from "../../../../../shared/models/IUser";
+import {ILecture} from "../../../../../shared/models/ILecture";
+
+export class Course implements ICourse{
+  _id: any;
+  active: boolean;
+  courseAdmin: IUser;
+  students: IUser[];
+  lectures: ILecture[];
   name: string;
   description: string;
   teacher: string;
