+++ shared/models/ILecture.ts
